requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
beautifulsoup4
pydantic
langsmith
langchain
sentence-transformers
groq
tenacity
aiolimiter
gtts
elevenlabs
